Reinitialize the locator when called on a different Plone Site.
Fixes https://github.com/collective/collective.mailchimp/issues/31.
Until now the registry and settings of the locator would be loaded once, the first time that a Plone Site with collective.mailchimp installed accesses it.
When you then install the add-on in a second Plone Site in the same Zope instance, they would share the registry and settings.
In other words:

- Go to the mailchimp control panel of site 1.
- The details under "MailChimp account" are from the api key of this site.
- Go to the mailchimp control panel of site 2.
- The details under "MailChimp account" are from the api key of the first site.

The reinitializing seems the least impactful way to fix this.
The alternative would be to add the locator as persistent utility in each site on install, and I don't like that.