Split the controlpanel into two separate pages: settings and data.
Until now, if the settings resulted in an error while fetching the data from mailchimp,
you could not fix this because the whole control panel would fail to load.
Fixes https://github.com/collective/collective.mailchimp/issues/26.

The control panel with the settings now has a link to new `@@mailchimp-data`, and this has a link back to the mailchimp control panel.